+++ lib/src/style.dart
@@ -1,6 +1,194 @@
 part of 'package:animated_toggle_switch/animated_toggle_switch.dart';
 
-class ToggleStyle {
+/// The base class for all toggle styles.
+abstract class _BaseToggleStyle {
+  const _BaseToggleStyle._();
+
+  ToggleStyleProperty<Color>? get _indicatorColor;
+
+  ToggleStyleProperty<Gradient>? get _indicatorGradient;
+
+  ToggleStyleProperty<Color>? get _backgroundColor;
+
+  ToggleStyleProperty<Gradient>? get _backgroundGradient;
+
+  ToggleStyleProperty<Color>? get _borderColor;
+
+  ToggleStyleProperty<BorderRadiusGeometry>? get _borderRadius;
+
+  ToggleStyleProperty<BorderRadiusGeometry>? get _indicatorBorderRadius;
+
+  ToggleStyleProperty<BoxBorder>? get _indicatorBorder;
+
+  ToggleStyleProperty<List<BoxShadow>>? get _indicatorBoxShadow;
+
+  ToggleStyleProperty<List<BoxShadow>>? get _boxShadow;
+
+  _BaseToggleStyle _merge(
+    _BaseToggleStyle? other,
+    BorderRadiusGeometry indicatorBorderRadiusDifference,
+  ) =>
+      other == null
+          ? this
+          : _CustomToggleStyle._(
+              indicatorColor: other._indicatorColor ?? _indicatorColor,
+              indicatorGradient: other._indicatorGradient ??
+                  (other._indicatorColor != null ? null : _indicatorGradient),
+              backgroundColor: other._backgroundColor ?? _backgroundColor,
+              backgroundGradient: other._backgroundGradient ??
+                  (other._backgroundColor != null ? null : _backgroundGradient),
+              borderColor: other._borderColor ?? _borderColor,
+              borderRadius: other._borderRadius ?? _borderRadius,
+              indicatorBorderRadius: other._indicatorBorderRadius ??
+                  other._borderRadius?._map((value) =>
+                      value.subtract(indicatorBorderRadiusDifference)) ??
+                  _indicatorBorderRadius ??
+                  _borderRadius?._map((value) =>
+                      value.subtract(indicatorBorderRadiusDifference)),
+              indicatorBorder: other._indicatorBorder ?? _indicatorBorder,
+              indicatorBoxShadow:
+                  other._indicatorBoxShadow ?? _indicatorBoxShadow,
+              boxShadow: other._boxShadow ?? _boxShadow,
+            );
+
+  static _BaseToggleStyle Function(
+      _BaseToggleStyle style1, _BaseToggleStyle style2, double t) _lerpFunction(
+          AnimationType animationType) =>
+      (style1, style2, t) => _lerp(style1, style2, t, animationType);
+
+  static _BaseToggleStyle _lerp(_BaseToggleStyle style1,
+          _BaseToggleStyle style2, double t, AnimationType animationType) =>
+      _CustomToggleStyle._(
+        indicatorColor: ToggleStyleProperty._lerpConditional(
+            style1._indicatorColor,
+            style2._indicatorColor,
+            t,
+            Color.lerp,
+            animationType),
+        indicatorGradient: ToggleStyleProperty._lerpConditional(
+            style1._indicatorGradient ??
+                style1._indicatorColor?._map((value) => value.toGradient()),
+            style2._indicatorGradient ??
+                style2._indicatorColor?._map((value) => value.toGradient()),
+            t,
+            Gradient.lerp,
+            animationType),
+        backgroundColor: ToggleStyleProperty._lerpConditional(
+            style1._backgroundColor,
+            style2._backgroundColor,
+            t,
+            Color.lerp,
+            animationType),
+        backgroundGradient: ToggleStyleProperty._lerpConditional(
+            style1._backgroundGradient ??
+                style1._backgroundColor?._map((value) => value.toGradient()),
+            style2._backgroundGradient ??
+                style2._backgroundColor?._map((value) => value.toGradient()),
+            t,
+            Gradient.lerp,
+            animationType),
+        borderColor: ToggleStyleProperty._lerpConditional(style1._borderColor,
+            style2._borderColor, t, Color.lerp, animationType),
+        borderRadius: ToggleStyleProperty._lerpConditional(style1._borderRadius,
+            style2._borderRadius, t, BorderRadiusGeometry.lerp, animationType),
+        indicatorBorderRadius: ToggleStyleProperty._lerpConditional(
+            style1._indicatorBorderRadius ?? style1._borderRadius,
+            style2._indicatorBorderRadius ?? style2._borderRadius,
+            t,
+            BorderRadiusGeometry.lerp,
+            animationType),
+        indicatorBorder: ToggleStyleProperty._lerpConditional(
+            style1._indicatorBorder,
+            style2._indicatorBorder,
+            t,
+            BoxBorder.lerp,
+            animationType),
+        indicatorBoxShadow: ToggleStyleProperty._lerpConditional(
+            style1._indicatorBoxShadow,
+            style2._indicatorBoxShadow,
+            t,
+            BoxShadow.lerpList,
+            animationType),
+        boxShadow: ToggleStyleProperty._lerpConditional(style1._boxShadow,
+            style2._boxShadow, t, BoxShadow.lerpList, animationType),
+      );
+}
+
+/// Currently not supported.
+class _CustomToggleStyle extends _BaseToggleStyle {
+  @override
+  final ToggleStyleProperty<Color>? _indicatorColor;
+  @override
+  final ToggleStyleProperty<Gradient>? _indicatorGradient;
+  @override
+  final ToggleStyleProperty<Color>? _backgroundColor;
+  @override
+  final ToggleStyleProperty<Gradient>? _backgroundGradient;
+  @override
+  final ToggleStyleProperty<Color>? _borderColor;
+  @override
+  final ToggleStyleProperty<BorderRadiusGeometry>? _borderRadius;
+  @override
+  final ToggleStyleProperty<BorderRadiusGeometry>? _indicatorBorderRadius;
+  @override
+  final ToggleStyleProperty<BoxBorder>? _indicatorBorder;
+  @override
+  final ToggleStyleProperty<List<BoxShadow>>? _indicatorBoxShadow;
+  @override
+  final ToggleStyleProperty<List<BoxShadow>>? _boxShadow;
+
+  /// Default constructor for [_CustomToggleStyle].
+  ///
+  /// If you don't want to disable the animation of single properties,
+  /// you should use [ToggleStyle] instead.
+  const _CustomToggleStyle({
+    ToggleStyleProperty<Color>? indicatorColor,
+    ToggleStyleProperty<Gradient>? indicatorGradient,
+    ToggleStyleProperty<Color>? backgroundColor,
+    ToggleStyleProperty<Gradient>? backgroundGradient,
+    ToggleStyleProperty<Color>? borderColor,
+    ToggleStyleProperty<BorderRadiusGeometry>? borderRadius,
+    ToggleStyleProperty<BorderRadiusGeometry>? indicatorBorderRadius,
+    ToggleStyleProperty<BoxBorder>? indicatorBorder,
+    ToggleStyleProperty<List<BoxShadow>>? indicatorBoxShadow,
+    ToggleStyleProperty<List<BoxShadow>>? boxShadow,
+  })  : _indicatorColor = indicatorColor,
+        _indicatorGradient = indicatorGradient,
+        _backgroundColor = backgroundColor,
+        _backgroundGradient = backgroundGradient,
+        _borderColor = borderColor,
+        _borderRadius = borderRadius,
+        _indicatorBorderRadius = indicatorBorderRadius,
+        _indicatorBorder = indicatorBorder,
+        _indicatorBoxShadow = indicatorBoxShadow,
+        _boxShadow = boxShadow,
+        super._();
+
+  const _CustomToggleStyle._({
+    required ToggleStyleProperty<Color>? indicatorColor,
+    required ToggleStyleProperty<Gradient>? indicatorGradient,
+    required ToggleStyleProperty<Color>? backgroundColor,
+    required ToggleStyleProperty<Gradient>? backgroundGradient,
+    required ToggleStyleProperty<Color>? borderColor,
+    required ToggleStyleProperty<BorderRadiusGeometry>? borderRadius,
+    required ToggleStyleProperty<BorderRadiusGeometry>? indicatorBorderRadius,
+    required ToggleStyleProperty<BoxBorder>? indicatorBorder,
+    required ToggleStyleProperty<List<BoxShadow>>? indicatorBoxShadow,
+    required ToggleStyleProperty<List<BoxShadow>>? boxShadow,
+  })  : _indicatorColor = indicatorColor,
+        _indicatorGradient = indicatorGradient,
+        _backgroundColor = backgroundColor,
+        _backgroundGradient = backgroundGradient,
+        _borderColor = borderColor,
+        _borderRadius = borderRadius,
+        _indicatorBorderRadius = indicatorBorderRadius,
+        _indicatorBorder = indicatorBorder,
+        _indicatorBoxShadow = indicatorBoxShadow,
+        _boxShadow = boxShadow,
+        super._();
+}
+
+class ToggleStyle extends _BaseToggleStyle {
   /// Background color of the indicator.
   ///
   /// Defaults to [ThemeData.colorScheme.secondary].
@@ -42,6 +230,9 @@ class ToggleStyle {
   final List<BoxShadow>? boxShadow;
 
   /// Default constructor for [ToggleStyle].
+  ///
+  /// If you want to disable the animation of single properties,
+  /// you can use [_CustomToggleStyle] for this purpose.
   const ToggleStyle({
     this.indicatorColor,
     this.indicatorGradient,
@@ -53,9 +244,8 @@ class ToggleStyle {
     this.indicatorBorder,
     this.indicatorBoxShadow,
     this.boxShadow,
-  });
+  }) : super._();
 
-  /// Private constructor for setting all possible parameters.
   const ToggleStyle._({
     required this.indicatorColor,
     required this.indicatorGradient,
@@ -67,78 +257,47 @@ class ToggleStyle {
     required this.indicatorBorder,
     required this.indicatorBoxShadow,
     required this.boxShadow,
-  });
+  }) : super._();
 
-  ToggleStyle _merge(
-    ToggleStyle? other,
-    BorderRadiusGeometry indicatorBorderRadiusDifference,
-  ) =>
-      other == null
-          ? this
-          : ToggleStyle._(
-              indicatorColor: other.indicatorColor ?? indicatorColor,
-              indicatorGradient: other.indicatorGradient ??
-                  (other.indicatorColor != null ? null : indicatorGradient),
-              backgroundColor: other.backgroundColor ?? backgroundColor,
-              backgroundGradient: other.backgroundGradient ??
-                  (other.backgroundColor != null ? null : backgroundGradient),
-              borderColor: other.borderColor ?? borderColor,
-              borderRadius: other.borderRadius ?? borderRadius,
-              indicatorBorderRadius: other.indicatorBorderRadius ??
-                  other.borderRadius
-                      .subtractNullable(indicatorBorderRadiusDifference) ??
-                  indicatorBorderRadius ??
-                  borderRadius
-                      .subtractNullable(indicatorBorderRadiusDifference),
-              indicatorBorder: other.indicatorBorder ?? indicatorBorder,
-              indicatorBoxShadow:
-                  other.indicatorBoxShadow ?? indicatorBoxShadow,
-              boxShadow: other.boxShadow ?? boxShadow,
-            );
+  @override
+  ToggleStyleProperty<Color>? get _backgroundColor =>
+      ToggleStyleProperty.nullable(backgroundColor);
 
-  static ToggleStyle _lerp(ToggleStyle style1, ToggleStyle style2, double t) =>
-      ToggleStyle._(
-        indicatorColor:
-            Color.lerp(style1.indicatorColor, style2.indicatorColor, t),
-        indicatorGradient: Gradient.lerp(
-          style1.indicatorGradient ?? style1.indicatorColor?.toGradient(),
-          style2.indicatorGradient ?? style2.indicatorColor?.toGradient(),
-          t,
-        ),
-        backgroundColor:
-            Color.lerp(style1.backgroundColor, style2.backgroundColor, t),
-        backgroundGradient: Gradient.lerp(
-          style1.backgroundGradient ?? style1.backgroundColor?.toGradient(),
-          style2.backgroundGradient ?? style2.backgroundColor?.toGradient(),
-          t,
-        ),
-        borderColor: Color.lerp(style1.borderColor, style2.borderColor, t),
-        borderRadius: BorderRadiusGeometry.lerp(
-          style1.borderRadius,
-          style2.borderRadius,
-          t,
-        ),
-        indicatorBorderRadius: BorderRadiusGeometry.lerp(
-          style1.indicatorBorderRadius ?? style1.borderRadius,
-          style2.indicatorBorderRadius ?? style2.borderRadius,
-          t,
-        ),
-        indicatorBorder: BoxBorder.lerp(
-          style1.indicatorBorder,
-          style2.indicatorBorder,
-          t,
-        ),
-        indicatorBoxShadow: BoxShadow.lerpList(
-          style1.indicatorBoxShadow,
-          style2.indicatorBoxShadow,
-          t,
-        ),
-        boxShadow: BoxShadow.lerpList(
-          style1.boxShadow,
-          style2.boxShadow,
-          t,
-        ),
-      );
+  @override
+  ToggleStyleProperty<Gradient>? get _backgroundGradient =>
+      ToggleStyleProperty.nullable(backgroundGradient);
+
+  @override
+  ToggleStyleProperty<Color>? get _borderColor =>
+      ToggleStyleProperty.nullable(borderColor);
+
+  @override
+  ToggleStyleProperty<BorderRadiusGeometry>? get _borderRadius =>
+      ToggleStyleProperty.nullable(borderRadius);
+
+  @override
+  ToggleStyleProperty<List<BoxShadow>>? get _boxShadow =>
+      ToggleStyleProperty.nullable(boxShadow);
+
+  @override
+  ToggleStyleProperty<BoxBorder>? get _indicatorBorder =>
+      ToggleStyleProperty.nullable(indicatorBorder);
+
+  @override
+  ToggleStyleProperty<BorderRadiusGeometry>? get _indicatorBorderRadius =>
+      ToggleStyleProperty.nullable(indicatorBorderRadius);
+
+  @override
+  ToggleStyleProperty<List<BoxShadow>>? get _indicatorBoxShadow =>
+      ToggleStyleProperty.nullable(indicatorBoxShadow);
+
+  @override
+  ToggleStyleProperty<Color>? get _indicatorColor =>
+      ToggleStyleProperty.nullable(indicatorColor);
+
+  @override
+  ToggleStyleProperty<Gradient>? get _indicatorGradient =>
+      ToggleStyleProperty.nullable(indicatorGradient);
 
   // coverage:ignore-start
 
@@ -174,9 +333,45 @@ class ToggleStyle {
 // coverage:ignore-end
 }
 
-extension _XSubtractNullable on BorderRadiusGeometry? {
-  BorderRadiusGeometry? subtractNullable(BorderRadiusGeometry other) {
-    if (this == null) return null;
-    return this!.subtract(other);
+class ToggleStyleProperty<T> {
+  /// The value of this property.
+  final T value;
+
+  /// Indicates if this property will be animated when changed.
+  final bool animationEnabled;
+
+  const ToggleStyleProperty(
+    this.value, {
+    this.animationEnabled = true,
+  });
+
+  static ToggleStyleProperty<T>? nullable<T>(
+    T? value, {
+    bool animationEnabled = true,
+  }) =>
+      value == null
+          ? null
+          : ToggleStyleProperty(value, animationEnabled: animationEnabled);
+
+  ToggleStyleProperty<S> _map<S>(S Function(T value) map) =>
+      ToggleStyleProperty(
+        map(value),
+        animationEnabled: animationEnabled,
+      );
+
+  static ToggleStyleProperty<T>? _lerpConditional<T>(
+      ToggleStyleProperty<T>? prop1,
+      ToggleStyleProperty<T>? prop2,
+      double t,
+      T? Function(T?, T?, double) lerp,
+      AnimationType animationType) {
+    if (prop1?.animationEnabled != true && prop2?.animationEnabled != true) {
+      if (animationType == AnimationType.onHover && t < 0.5) return prop1;
+      return prop2;
+    }
+    return ToggleStyleProperty.nullable(
+      lerp(prop1?.value, prop2?.value, t),
+      animationEnabled: true,
+    );
   }
 }